+++ apps/frontend/src/components/organisms/roleManagement/PendingApplications.tsx
@@ -11,39 +11,7 @@ import {
 import { application } from "@/components/molecules/roleManagement/PendingApplicationsColumns";
 
 async function getData(): Promise<application[]> {
-  return [
-    {
-      id: "3b8d9f1a",
-      name: "Robert Wilson",
-      email: "robert.wilson@up.ac.za",
-      appliedRole: "Lecturer",
-    },
-    {
-      id: "7c2e4a6b",
-      name: "Laura Martinez",
-      email: "laura.martinez@up.ac.za",
-      appliedRole: "Uni Admin",
-    },
-    {
-      id: "9f1a3b8d",
-      name: "David Kim",
-      email: "david.kim@up.ac.za",
-      appliedRole: "Student",
-    },
-    {
-      id: "2a6b7c2e",
-      name: "Anna Nguyen",
-      email: "anna.nguyen@up.ac.za",
-      appliedRole: "Lecturer",
-    },
-    {
-      id: "5d4e3f2a",
-      name: "Michael Brown",
-      email: "michael.brown@up.ac.za",
-      appliedRole: "Student",
-    },
-    // ...
-  ];
+  return [];
 }
 
 export default async function PendingApplicationsCard() {
